+++ src/pages/auth/privacy-policy/index.tsx
@@ -8,7 +8,7 @@ export default function TermsOfServicePage() {
 
   return (
     <TextContainer>
-      <BackToPrevious href={authNavigationUrls.signUp()} title="Back to sign up" />
+      <BackToPrevious href={authNavigationUrls.signUp()} title="Back to previous page" />
       <p>{t.privacyDescription.title}</p>
       <span>{t.privacyDescription.description}</span>
     </TextContainer>
