+++ src/ui/back-to-previous/backtoprevious.tsx
@@ -1,4 +1,5 @@
 import Link from 'next/link'
+import { useRouter } from 'next/router'
 
 import { BackToPreviousIcon } from '@/app/assets/svg'
 import s from '@/ui/back-to-previous/backtoprevious.module.scss'
@@ -8,9 +9,17 @@ type PropsType = {
   href: string
 }
 export const BackToPrevious = ({ title, href }: PropsType) => {
+  const router = useRouter()
+
   return (
     <div className={s.wrapper}>
-      <Link href={href}>
+      <Link
+        href={href}
+        onClick={e => {
+          e.preventDefault()
+          router.back()
+        }}
+      >
         <BackToPreviousIcon />
         <span>{title}</span>
       </Link>
